# Benchmark VISTA3D