### Description
This is a function to benchmark, analyze and optimize VISTA3D bundle to get a better inference latency. I will try to add all the benchmark results and analyses in the PR comments. Meanwhile the general conclusion will be updated here in the PR description.

### Status
**Work in progress**

### Conclusion
1. A larger sw_batch_size (~10) can optimize the latency on the A100(80GB) GPU. 

### TODO:
- [x] Add timing function to analyze the latency of each part of the bundle.
- [x] Find out the bottleneck of the latency.
- [ ] Analyze the detail latency of each part of the VISTA3D network.
- [ ] Compare the TRT model and non-TRT model.
- [ ] Optimize the latency with the TRT and other methods.