@binliunls 2024-09-23
As the `Range` function is added to the bundle, we can take a look at the latency detail of one inference iteration, which is shown below. All the gray boxes under the `SW_Patchforward_Loop` stand for a model computation/prediction with the VISTA3D network mentioned as `SW_Model_Computation` latter. Since the bundle uses the sliding window inference and the given image size(512x512x77) is larger than the sliding window size(128x128x128), one sliding window inference iteration actually contains several iterations of `SW_Model_Computation`. 

![image](https://github.com/user-attachments/assets/96818ff6-5794-48dd-aee6-9c886eb49d6e)

By zooming into the red box, we can further analyze the latency of one `SW_Model_Computation` iteration.  As shown in the image below, each `SW_Model_Computation` will excute the `cuStreamSynchronize (orange and green) --> cudaFree(pink) --> cuStreamSynchronize (orange and green)` in order. These overhead steps take a large percentage of `SW_Model_Computation`, which are introduced by the **loop inference**.

![image](https://github.com/user-attachments/assets/c7c94978-04cf-4a99-9bfd-ac0491679f5f)