@binliunls 2024-09-23
The relation of latency vs sw_batch_size is:

| sw_batch_size | 4 | 8 | 10 | 12 | 14 | 16 | 20 |
| :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
| latency | 2.803 | 2.508 | 2.362 | 2.289 | 2.301 | 2.594 | 2.74 |

![image](https://github.com/user-attachments/assets/a1fb37fc-de9c-4882-ae41-823d55769610)