@binliunls 2024-09-24
Set the `sw_batch_size` to 14 and run the bundle with TRT inference (only compile encoder and run all classes segmentation). The benchmarks are shown below. The upper one is the latency detail about the original bundle, while the lower one is TRT bundler.

Hi @borisfom  I didn't see a significant improvement about the encoder. The inference latency of TRT and non-TRT are nearly the same. Could you please offer some suggestion here? Thanks in advance!

**Original bundle**
![image](https://github.com/user-attachments/assets/22a43455-7f47-4a16-9359-0dd38cdf3dfe)

**TRT bundle**
![image](https://github.com/user-attachments/assets/1dbee00e-98af-494b-a429-b1d93fe17e25)