@binliunls 2024-09-26
> @binliunls : well that seems to be the case TRT does not help much with this net - I was not able to run batch=14 on my box but batch=8 gave similar result. In fact, batch=1 does not give a lot of improvement, too. It looks like TRT is currently only running a small fraction of encoder's forward action - and expanding that may not be straightforward. I will look a bit more at the model, but most likely, expanding converted part won't be possible as that would make it dependent on dynamic arguments.

This does improve the inference on the V100 32GB GPU with max batchsize 6. Here are the details.

MONAI bundle:
![image](https://github.com/user-attachments/assets/77fb0f1e-de3f-4350-8765-93833681d9c5)

TRT bundle:
![image](https://github.com/user-attachments/assets/c000ec03-4422-48ad-9163-ac5a6d3c1ce3)