@binliunls 2024-09-26
Note that there is a memory malloc that is unecessary for all classes inference, which is related to [this line](https://github.com/binliunls/MONAI/pull/1/files#r1777427703]

![image](https://github.com/user-attachments/assets/4682aca9-a4e2-473c-b51d-03d2af49bb2c)