@binliunls 2024-09-27
> @binliunls : wow, that's massive sync apparently caused by waiting for TRT results - does it actually help removing it though ?

Hi @borisfom , the result didn't use TRT. It's just a straightforward MONAI bundle. And yes, remove it will help to improve latency, since the removing of cudaMalloc has already reduce like 200-300ms. I will try to figure out where these API call happened in the code and see if we can further improve the performance. 

Thanks  
Bin