@binliunls 2024-09-27
The embedding mask part of the classification header is another high latency part that can be optimized, which is shown in the image below.

![image](https://github.com/user-attachments/assets/b58d3c6d-bd05-43ff-9fc4-4b56829816e2)

 It's using a for-loop in python to perform the tensor multiplication which is low efficiency.  The [code snippet](https://github.com/binliunls/MONAI/pull/1/files#r1778869153) is like:

```python
b, c, h, w, d = src.shape
masks = []
for i in range(b):
    mask = class_embedding @ src[[i]].view(1, c, h * w * d)
    masks.append(mask.view(-1, 1, h, w, d))
``` 

We can refactor it to do a broadcast tensor multiplication call like:

```python
b, c, h, w, d = src.shape
c = class_embedding.squeeze() @ src.view(b, c, h * w * d)
```

Here is a simple test case to test these two implement are the same:

```python
import torch

def mat_mul2(class_embedding, src):
    b, c, h, w, d = src.shape
    c = class_embedding.squeeze() @ src.view(b, c, h * w * d)
    c = c.view(b, -1, h, w, d)
    return torch.transpose(c, 0, 1)

def mat_mul1(class_embedding, src):
    b, c, h, w, d = src.shape
    masks = []
    for i in range(b):
        mask = class_embedding @ src[[i]].view(1, c, h * w * d)
        masks.append(mask.view(-1, 1, h, w, d))
    return torch.cat(masks, 1)

a = torch.rand((17, 1, 4))
b = torch.rand(4, 4, 12, 12, 12)
ans1 = mat_mul1(a, b)
ans2 = mat_mul2(a, b)
assert torch.allclose(ans1, ans2)
```

After replacing the embedding mask calculation with the new way, we can reduce the whole sliding inference latency from 2.302s to 1.830s. The embedding mask part latency reduced from 278ms to 4ms.

![image](https://github.com/user-attachments/assets/143e68d2-22fb-4015-9cb0-31830c6b736c)